requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
blinker==1.9.0
ccxt==4.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==45.0.7
Flask==3.1.2
frozenlist==1.7.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
keyboard==0.13.5
MarkupSafe==3.0.2
multidict==6.6.4
numpy==2.3.2
pandas==2.3.2
propcache==0.3.2
pycares==4.10.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
